Pass detectors view models through the environment
Using the environment is an alternative to passing a detectors view model as a parameter in an init from a parent view.

The approach we replace did not require us to have the view model as a property in the child view. The value was only used to initialize an error alert factory.